Move ScaleAddBuffer to a new file
