+++ resolve/workspace/LinAlgWorkspaceHIP.hpp
@@ -10,25 +10,6 @@
 namespace ReSolve
 {
 
-  class ScaleAddBufferHIP
-  {
-  public:
-    ScaleAddBufferHIP(index_type numRows);
-    ~ScaleAddBufferHIP();
-    index_type*         getRowData();
-    rocsparse_mat_descr getMatrixDescriptor();
-    index_type          getNumRows();
-    void                setNnz(index_type nnz);
-    index_type          getNnz();
-
-  private:
-    index_type*         rowData_;
-    rocsparse_mat_descr mat_A_;
-    index_type          numRows_;
-    index_type          nnz_;
-    MemoryHandler       mem_;
-  };
-
   class LinAlgWorkspaceHIP
   {
   public:
